[css-grid] Migrate grid-self-baseline-* tests to WPT
Migrate all these tests out of blink/web_tests/fast/css-grid-layout
and into the WPT-specific directory, adding links to the relevant
specs and a test assertion describing its purpose.

This is a fairly big CL because it migrates 28 tests at once, but
we decided to do that since the changes done for all those tests
are pretty much the same, so both the elaboration of this patch
and the review might become easier this way, as changes are quite
mechanical in general.

Bug: 767015, 1063749
Change-Id: Ib19287254bf6aa70553ba805b2844e36884f52e7