[PlayBilling] Total optional on only app-store billing requested
Context:
Total is a field in the "new PaymentRequest()" API. It specifies the
amount and currency of the payment request. However, when the merchant
requests for the app-store billing (e.g., Google Play Store billing -
by specifying "https://play.google.com/billing" as the method), the
total field becomes unnecessary. This is because app-stores takes
the total from elsewhere.

Before:
The total field is mandatory for PaymentRequest.

After:
The total field is optional if only app-store methods are requested.
When total field is optional and left out, Chrome would add a total of
amount 0, currency "PBM" and label "AppStoreBillingTotalPlaceHolder".

Bug: 1066531

Change-Id: Id5ad87b9fc452fd41a1ebef066d981737545a235